Remove delegate cache functionality outside of the blockchain abstraction layer
